+++ lib/activesupport/logger.rb
@@ -66,11 +66,3 @@ module ActiveSupport
     private_class_method :normalize_sources
   end
 end
-
-require "version_gem"
-
-ActiveSupport::Logger::Version.class_eval do
-  extend VersionGem::Basic
-end
-
-require_relative "logger/version"